test: Moved failed connection test into integration tests; the explicitly don't require OBS, and it may work better in cicd
